+++ src/frontend/SourceManager.cpp
@@ -166,6 +166,14 @@ namespace panther{
 		// String (4)
 		this->base_types.emplace_back(PIR::BaseType::Kind::Builtin, Token::TypeString);
 		this->types.emplace_back( PIR::Type(PIR::BaseType::ID(4)) );
+
+		// ISize (5)
+		this->base_types.emplace_back(PIR::BaseType::Kind::Builtin, Token::TypeISize);
+		this->types.emplace_back( PIR::Type(PIR::BaseType::ID(5)) );
+
+		// USize (6)
+		this->base_types.emplace_back(PIR::BaseType::Kind::Builtin, Token::TypeUSize);
+		this->types.emplace_back( PIR::Type(PIR::BaseType::ID(6)) );
 	};
 
 
@@ -175,135 +183,70 @@ namespace panther{
 		using ParamKind = AST::FuncParams::Param::Kind;
 
 
+		auto create_func_base_type = [&](evo::ArrayProxy<PIR::BaseType::Operator::Param> params, PIR::Type::ID return_type) noexcept -> PIR::BaseType::ID {
+			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
+			base_type.callOperator = PIR::BaseType::Operator(
+				std::vector<PIR::BaseType::Operator::Param>(params.begin(), params.end()),
+				return_type
+			);
+
+			return this->createBaseType(std::move(base_type));
+		};
+
+
 		///////////////////////////////////
 		// setup protos
 
-		const PIR::BaseType::ID math_ops_Int_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeInt(), ParamKind::Read}, {this->getTypeInt(), ParamKind::Read}, },
-				this->getTypeInt()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
-
-		const PIR::BaseType::ID math_ops_UInt_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeUInt(), ParamKind::Read}, {this->getTypeUInt(), ParamKind::Read}, },
-				this->getTypeUInt()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
+		const PIR::BaseType::ID math_ops_Int_type = create_func_base_type({{this->getTypeInt(), ParamKind::Read}, {this->getTypeInt(), ParamKind::Read}}, this->getTypeInt());
+		const PIR::BaseType::ID math_ops_UInt_type = create_func_base_type({{this->getTypeUInt(), ParamKind::Read}, {this->getTypeUInt(), ParamKind::Read}}, this->getTypeUInt());
+		const PIR::BaseType::ID math_ops_ISize_type = create_func_base_type({{this->getTypeISize(), ParamKind::Read}, {this->getTypeISize(), ParamKind::Read}}, this->getTypeISize());
+		const PIR::BaseType::ID math_ops_USize_type = create_func_base_type({{this->getTypeUSize(), ParamKind::Read}, {this->getTypeUSize(), ParamKind::Read}}, this->getTypeUSize());
 
 
-		const PIR::BaseType::ID logical_Int_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeInt(), ParamKind::Read}, {this->getTypeInt(), ParamKind::Read}, },
-				this->getTypeBool()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
-		const PIR::BaseType::ID logical_UInt_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeUInt(), ParamKind::Read}, {this->getTypeUInt(), ParamKind::Read}, },
-				this->getTypeBool()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
-		const PIR::BaseType::ID logical_Bool_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeBool(), ParamKind::Read}, {this->getTypeBool(), ParamKind::Read}, },
-				this->getTypeBool()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
+		const PIR::BaseType::ID logical_Int_type = create_func_base_type({{this->getTypeInt(), ParamKind::Read}, {this->getTypeInt(), ParamKind::Read}}, this->getTypeBool());
+		const PIR::BaseType::ID logical_UInt_type = create_func_base_type({{this->getTypeUInt(), ParamKind::Read}, {this->getTypeUInt(), ParamKind::Read}}, this->getTypeBool());
+		const PIR::BaseType::ID logical_Bool_type = create_func_base_type({{this->getTypeBool(), ParamKind::Read}, {this->getTypeBool(), ParamKind::Read}}, this->getTypeBool());
+		const PIR::BaseType::ID logical_ISize_type = create_func_base_type({{this->getTypeISize(), ParamKind::Read}, {this->getTypeISize(), ParamKind::Read}}, this->getTypeBool());
+		const PIR::BaseType::ID logical_USize_type = create_func_base_type({{this->getTypeUSize(), ParamKind::Read}, {this->getTypeUSize(), ParamKind::Read}}, this->getTypeBool());
 
 
-		const PIR::BaseType::ID Int_in_Int_out_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeInt(), ParamKind::Read} },
-				this->getTypeInt()
-			);
+		const PIR::BaseType::ID Int_in_Int_out_type = create_func_base_type({ {this->getTypeInt(), ParamKind::Read} }, this->getTypeInt());
+		const PIR::BaseType::ID ISize_in_ISize_out_type = create_func_base_type({ {this->getTypeISize(), ParamKind::Read} }, this->getTypeISize());
+		const PIR::BaseType::ID Bool_in_Bool_out_type = create_func_base_type({ {this->getTypeBool(), ParamKind::Read} }, this->getTypeBool());
+		
 
-			return this->createBaseType(std::move(base_type));
-		}();
-		const PIR::BaseType::ID Bool_in_Bool_out_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeBool(), ParamKind::Read} },
-				this->getTypeBool()
-			);
+		// Int
+		const PIR::BaseType::ID Int_in_UInt_out_type = create_func_base_type({ {this->getTypeInt(), ParamKind::Read} }, this->getTypeUInt());
+		const PIR::BaseType::ID Int_in_Bool_out_type = create_func_base_type({ {this->getTypeInt(), ParamKind::Read} }, this->getTypeBool());
+		const PIR::BaseType::ID Int_in_ISize_out_type = create_func_base_type({ {this->getTypeInt(), ParamKind::Read} }, this->getTypeISize());
+		const PIR::BaseType::ID Int_in_USize_out_type = create_func_base_type({ {this->getTypeInt(), ParamKind::Read} }, this->getTypeUSize());
 
-			return this->createBaseType(std::move(base_type));
-		}();
+		// UInt
+		const PIR::BaseType::ID UInt_in_Int_out_type = create_func_base_type({ {this->getTypeUInt(), ParamKind::Read} }, this->getTypeInt());
+		const PIR::BaseType::ID UInt_in_Bool_out_type = create_func_base_type({ {this->getTypeUInt(), ParamKind::Read} }, this->getTypeBool());
+		const PIR::BaseType::ID UInt_in_ISize_out_type = create_func_base_type({ {this->getTypeUInt(), ParamKind::Read} }, this->getTypeISize());
+		const PIR::BaseType::ID UInt_in_USize_out_type = create_func_base_type({ {this->getTypeUInt(), ParamKind::Read} }, this->getTypeUSize());
+
+		// Bool
+		const PIR::BaseType::ID Bool_in_Int_out_type = create_func_base_type({ {this->getTypeBool(), ParamKind::Read} }, this->getTypeInt());
+		const PIR::BaseType::ID Bool_in_UInt_out_type = create_func_base_type({ {this->getTypeBool(), ParamKind::Read} }, this->getTypeUInt());
+		const PIR::BaseType::ID Bool_in_ISize_out_type = create_func_base_type({ {this->getTypeBool(), ParamKind::Read} }, this->getTypeISize());
+		const PIR::BaseType::ID Bool_in_USize_out_type = create_func_base_type({ {this->getTypeBool(), ParamKind::Read} }, this->getTypeUSize());
+
+		// ISize
+		const PIR::BaseType::ID ISize_in_Int_out_type = create_func_base_type({ {this->getTypeISize(), ParamKind::Read} }, this->getTypeInt());
+		const PIR::BaseType::ID ISize_in_UInt_out_type = create_func_base_type({ {this->getTypeISize(), ParamKind::Read} }, this->getTypeUInt());
+		const PIR::BaseType::ID ISize_in_Bool_out_type = create_func_base_type({ {this->getTypeISize(), ParamKind::Read} }, this->getTypeBool());
+		const PIR::BaseType::ID ISize_in_USize_out_type = create_func_base_type({ {this->getTypeISize(), ParamKind::Read} }, this->getTypeUSize());
+
+		// USize
+		const PIR::BaseType::ID USize_in_Int_out_type = create_func_base_type({ {this->getTypeUSize(), ParamKind::Read} }, this->getTypeInt());
+		const PIR::BaseType::ID USize_in_UInt_out_type = create_func_base_type({ {this->getTypeUSize(), ParamKind::Read} }, this->getTypeUInt());
+		const PIR::BaseType::ID USize_in_Bool_out_type = create_func_base_type({ {this->getTypeUSize(), ParamKind::Read} }, this->getTypeBool());
+		const PIR::BaseType::ID USize_in_ISize_out_type = create_func_base_type({ {this->getTypeUSize(), ParamKind::Read} }, this->getTypeISize());
 
 
-		const PIR::BaseType::ID Int_in_UInt_out_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeInt(), ParamKind::Read} },
-				this->getTypeUInt()
-			);
 
-			return this->createBaseType(std::move(base_type));
-		}();
-		const PIR::BaseType::ID Int_in_Bool_out_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeInt(), ParamKind::Read} },
-				this->getTypeBool()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
-
-		const PIR::BaseType::ID UInt_in_Int_out_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeUInt(), ParamKind::Read} },
-				this->getTypeInt()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
-		const PIR::BaseType::ID UInt_in_Bool_out_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeUInt(), ParamKind::Read} },
-				this->getTypeBool()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
-
-		const PIR::BaseType::ID Bool_in_Int_out_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeBool(), ParamKind::Read} },
-				this->getTypeInt()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
-		const PIR::BaseType::ID Bool_in_UInt_out_type = [&]() noexcept {
-			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
-			base_type.callOperator = PIR::BaseType::Operator(
-				std::vector<PIR::BaseType::Operator::Param>{ {this->getTypeBool(), ParamKind::Read} },
-				this->getTypeUInt()
-			);
-
-			return this->createBaseType(std::move(base_type));
-		}();
 		
 
 
@@ -351,31 +294,46 @@ namespace panther{
 
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::addInt, "addInt", math_ops_Int_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::addUInt, "addUInt", math_ops_UInt_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::addISize, "addISize", math_ops_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::addUSize, "addUSize", math_ops_USize_type);
 
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::addWrapInt, "addWrapInt", math_ops_Int_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::addWrapUInt, "addWrapUInt", math_ops_UInt_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::addWrapISize, "addWrapISize", math_ops_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::addWrapUSize, "addWrapUSize", math_ops_USize_type);
 
 
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::subInt, "subInt", math_ops_Int_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::subUInt, "subUInt", math_ops_UInt_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::subISize, "subISize", math_ops_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::subUSize, "subUSize", math_ops_USize_type);
 
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::subWrapInt, "subWrapInt", math_ops_Int_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::subWrapUInt, "subWrapUInt", math_ops_UInt_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::subWrapISize, "subWrapISize", math_ops_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::subWrapUSize, "subWrapUSize", math_ops_USize_type);
 
 
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::mulInt, "mulInt", math_ops_Int_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::mulUInt, "mulUInt", math_ops_UInt_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::mulISize, "mulISize", math_ops_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::mulUSize, "mulUSize", math_ops_USize_type);
 		
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::mulWrapInt, "mulWrapInt", math_ops_Int_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::mulWrapUInt, "mulWrapUInt", math_ops_UInt_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::mulWrapISize, "mulWrapISize", math_ops_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::mulWrapUSize, "mulWrapUSize", math_ops_USize_type);
 
 
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::divInt, "divInt", math_ops_Int_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::divUInt, "divUInt", math_ops_UInt_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::divISize, "divISize", math_ops_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::divUSize, "divUSize", math_ops_USize_type);
 
 
 
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::negateInt, "negateInt", Int_in_Int_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::negateISize, "negateISize", ISize_in_ISize_out_type);
 
 
 		///////////////////////////////////
@@ -388,7 +346,6 @@ namespace panther{
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::greaterThanInt, "greaterThanInt", logical_Int_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::greaterThanEqualInt, "greaterThanEqualInt", logical_Int_type);
 
-
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::equalUInt, "equalUInt", logical_UInt_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::notEqualUInt, "notEqualUInt", logical_UInt_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::lessThanUInt, "lessThanUInt", logical_UInt_type);
@@ -398,10 +355,23 @@ namespace panther{
 
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::equalBool, "equalBool", logical_Bool_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::notEqualBool, "notEqualBool", logical_Bool_type);
-		
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::logicalAnd, "logicalAnd", logical_Bool_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::logicalOr, "logicalOr", logical_Bool_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::logicalNot, "logicalNot", Bool_in_Bool_out_type);
+
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::equalISize, "equalISize", logical_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::notEqualISize, "notEqualISize", logical_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::lessThanISize, "lessThanISize", logical_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::lessThanEqualISize, "lessThanEqualISize", logical_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::greaterThanISize, "greaterThanISize", logical_ISize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::greaterThanEqualISize, "greaterThanEqualISize", logical_ISize_type);
+
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::equalUSize, "equalUSize", logical_USize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::notEqualUSize, "notEqualUSize", logical_USize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::lessThanUSize, "lessThanUSize", logical_USize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::lessThanEqualUSize, "lessThanEqualUSize", logical_USize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::greaterThanUSize, "greaterThanUSize", logical_USize_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::greaterThanEqualUSize, "greaterThanEqualUSize", logical_USize_type);
 
 
 		///////////////////////////////////
@@ -409,10 +379,28 @@ namespace panther{
 
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convIntToUInt, "convIntToUInt", Int_in_UInt_out_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convIntToBool, "convIntToBool", Int_in_Bool_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convIntToISize, "convIntToISize", Int_in_ISize_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convIntToUSize, "convIntToUSize", Int_in_USize_out_type);
+
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convUIntToInt, "convUIntToInt", UInt_in_Int_out_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convUIntToBool, "convUIntToBool", UInt_in_Bool_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convUIntToISize, "convUIntToISize", UInt_in_ISize_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convUIntToUSize, "convUIntToUSize", UInt_in_USize_out_type);
+
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convBoolToInt, "convBoolToInt", Bool_in_Int_out_type);
 		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convBoolToUInt, "convBoolToUInt", Bool_in_UInt_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convBoolToISize, "convBoolToISize", Bool_in_ISize_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convBoolToUSize, "convBoolToUSize", Bool_in_USize_out_type);
+
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convISizeToInt, "convISizeToInt", ISize_in_Int_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convISizeToUInt, "convISizeToUInt", ISize_in_UInt_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convISizeToBool, "convISizeToBool", ISize_in_Bool_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convISizeToUSize, "convISizeToUSize", ISize_in_USize_out_type);
+
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convUSizeToInt, "convUSizeToInt", USize_in_Int_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convUSizeToUInt, "convUSizeToUInt", USize_in_UInt_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convUSizeToBool, "convUSizeToBool", USize_in_Bool_out_type);
+		this->intrinsics.emplace_back(PIR::Intrinsic::Kind::convUSizeToISize, "convUSizeToISize", USize_in_ISize_out_type);
 
 
 
@@ -467,7 +455,11 @@ namespace panther{
 			for(uint32_t i = 0; i < uint32_t(PIR::Intrinsic::Kind::_MAX_); i+=1){
 				const PIR::Intrinsic::Kind i_kind = static_cast<PIR::Intrinsic::Kind>(i);
 				const PIR::Intrinsic& intrinsic = this->getIntrinsic(i_kind);
-				evo::Assert(intrinsic.kind == i_kind, "found an intrinsic with incorrect cooresponding kind");
+
+				evo::Assert(
+					i_kind == intrinsic.kind,
+					std::format("found an intrinsic with incorrect cooresponding kind (kind: {} != intrinsic: {})", int(i_kind), int(intrinsic.kind))
+				);
 			}
 		#endif
 
@@ -478,6 +470,8 @@ namespace panther{
 		PIR::BaseType& type_Int = this->getBaseType(Token::TypeInt);
 		PIR::BaseType& type_UInt = this->getBaseType(Token::TypeUInt);
 		PIR::BaseType& type_Bool = this->getBaseType(Token::TypeBool);
+		PIR::BaseType& type_ISize = this->getBaseType(Token::TypeISize);
+		PIR::BaseType& type_USize = this->getBaseType(Token::TypeUSize);
 
 
 		///////////////////////////////////
@@ -485,30 +479,45 @@ namespace panther{
 
 		type_Int.ops.add.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::addInt));
 		type_UInt.ops.add.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::addUInt));
+		type_ISize.ops.add.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::addISize));
+		type_USize.ops.add.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::addUSize));
 
 		type_Int.ops.addWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::addWrapInt));
 		type_UInt.ops.addWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::addWrapUInt));
+		type_ISize.ops.addWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::addWrapISize));
+		type_USize.ops.addWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::addWrapUSize));
 
 
 		type_Int.ops.sub.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::subInt));
 		type_UInt.ops.sub.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::subUInt));
+		type_ISize.ops.sub.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::subISize));
+		type_USize.ops.sub.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::subUSize));
 
 		type_Int.ops.subWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::subWrapInt));
 		type_UInt.ops.subWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::subWrapUInt));
+		type_ISize.ops.subWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::subWrapISize));
+		type_USize.ops.subWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::subWrapUSize));
 
 
 		type_Int.ops.mul.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::mulInt));
 		type_UInt.ops.mul.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::mulUInt));
+		type_ISize.ops.mul.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::mulISize));
+		type_USize.ops.mul.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::mulUSize));
 
 		type_Int.ops.mulWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::mulWrapInt));
 		type_UInt.ops.mulWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::mulWrapUInt));
+		type_ISize.ops.mulWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::mulWrapISize));
+		type_USize.ops.mulWrap.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::mulWrapUSize));
 
 
 		type_Int.ops.div.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::divInt));
 		type_UInt.ops.div.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::divUInt));
+		type_ISize.ops.div.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::divISize));
+		type_USize.ops.div.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::divUSize));
 
 
 		type_Int.ops.negate.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::negateInt));
+		type_ISize.ops.negate.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::negateISize));
 
 
 		///////////////////////////////////
@@ -517,22 +526,36 @@ namespace panther{
 		type_Int.ops.logicalEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::equalInt));
 		type_UInt.ops.logicalEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::equalUInt));
 		type_Bool.ops.logicalEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::equalBool));
+		type_ISize.ops.logicalEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::equalISize));
+		type_USize.ops.logicalEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::equalUSize));
 
 		type_Int.ops.notEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::notEqualInt));
 		type_UInt.ops.notEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::notEqualUInt));
 		type_Bool.ops.notEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::notEqualBool));
+		type_ISize.ops.notEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::notEqualISize));
+		type_USize.ops.notEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::notEqualUSize));
 
 		type_Int.ops.lessThan.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::lessThanInt));
 		type_UInt.ops.lessThan.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::lessThanUInt));
+		type_ISize.ops.lessThan.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::lessThanISize));
+		type_USize.ops.lessThan.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::lessThanUSize));
 
 		type_Int.ops.lessThanEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::lessThanEqualInt));
 		type_UInt.ops.lessThanEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::lessThanEqualUInt));
+		type_ISize.ops.lessThanEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::lessThanEqualISize));
+		type_USize.ops.lessThanEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::lessThanEqualUSize));
+
+
 
 		type_Int.ops.greaterThan.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::greaterThanInt));
 		type_UInt.ops.greaterThan.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::greaterThanUInt));
+		type_ISize.ops.greaterThan.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::greaterThanISize));
+		type_USize.ops.greaterThan.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::greaterThanUSize));
 
 		type_Int.ops.greaterThanEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::greaterThanEqualInt));
 		type_UInt.ops.greaterThanEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::greaterThanEqualUInt));
+		type_ISize.ops.greaterThanEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::greaterThanEqualISize));
+		type_USize.ops.greaterThanEqual.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::greaterThanEqualUSize));
 
 
 		type_Bool.ops.logicalNot.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::logicalNot));
@@ -545,13 +568,28 @@ namespace panther{
 
 		type_Int.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convIntToUInt));
 		type_Int.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convIntToBool));
+		type_Int.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convIntToISize));
+		type_Int.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convIntToUSize));
 
 		type_UInt.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convUIntToInt));
 		type_UInt.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convUIntToBool));
+		type_UInt.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convUIntToISize));
+		type_UInt.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convUIntToUSize));
 
 		type_Bool.ops.as.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convBoolToInt));
 		type_Bool.ops.as.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convBoolToUInt));
+		type_Bool.ops.as.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convBoolToISize));
+		type_Bool.ops.as.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convBoolToUSize));
 
+		type_ISize.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convISizeToInt));
+		type_ISize.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convISizeToUInt));
+		type_ISize.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convISizeToBool));
+		type_ISize.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convISizeToUSize));
+
+		type_USize.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convUSizeToInt));
+		type_USize.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convUSizeToUInt));
+		type_USize.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convUSizeToBool));
+		type_USize.ops.cast.emplace_back(this->getIntrinsicID(PIR::Intrinsic::Kind::convUSizeToISize));
 	};